Identify client through cookie
When at a meeting, the connection to the server may be momentarily
lost:
- Network failure
- WebSocket connection lost
- iPhone application switch from the browser to another app
- Accidentally closing the browser tab

While the connection is lost the client tab should keep working with
the last known rate.

When connection is regained both the rate and the time reference should
be updated.

Inconsistency between multiple tabs in the same client can also be
avoided by checking the client id against the list of attendees on the
meeting's client model when syncing and updating the view accordingly.